AAPLOT_optimized/requirements.txt
# AAPlot 优化版本依赖包
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.9.0
PyYAML>=6.0
